requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

av==9.2.0
joblib==1.2.0
librosa==0.10.0.post2
matplotlib==3.5.1
MMSA-FET==0.3.0
moviepy==1.0.3
numpy==1.24.2
nvidia-ml-py3==7.352.0
pandas==2.1.0
PyYAML==6.0
resampy==0.4.2
scikit-learn==1.2.2
soundfile==0.12.1
SpeechRecognition==3.8.1
termcolor==2.1.1
torchinfo==1.8.0
tqdm==4.65.0
transformers==4.27.3
visualdl==2.5.3
